move MlInput abstraction to Solidago, renamed into TournesolInput
