+++ solidago/src/solidago/pipeline/__init__.py
@@ -1,0 +1,1 @@
+from .inputs import TournesolInput
